+++ ruma-client-api/src/r0/media/get_content.rs
@@ -1,7 +1,7 @@
 //! [GET /_matrix/media/r0/download/{serverName}/{mediaId}](https://matrix.org/docs/spec/client_server/r0.6.0#get-matrix-media-r0-download-servername-mediaid)
 
 use ruma_api::ruma_api;
-use ruma_identifiers::{MxcUri, ServerName};
+use ruma_identifiers::{Error, MxcUri, ServerName};
 
 ruma_api! {
     metadata: {
@@ -59,8 +59,12 @@ impl<'a> Request<'a> {
     }
 
     /// Creates a new `Request` with the given url.
-    pub fn from_url(url: &'a MxcUri) -> Self {
-        Self { media_id: url.media_id(), server_name: url.server_name(), allow_remote: true }
+    pub fn from_url(url: &'a MxcUri) -> Result<Self, Error> {
+        if let (Some(media_id), Some(server_name)) = (url.media_id(), url.server_name()) {
+            Ok(Self { media_id, server_name, allow_remote: true })
+        } else {
+            Err(Error::InvalidMxcUri)
+        }
     }
 }