+++ ruma-client-api/src/r0/media/get_content_thumbnail.rs
@@ -2,7 +2,7 @@
 
 use js_int::UInt;
 use ruma_api::ruma_api;
-use ruma_identifiers::{MxcUri, ServerName};
+use ruma_identifiers::{Error, MxcUri, ServerName};
 use ruma_serde::StringEnum;
 
 /// The desired resizing method.
@@ -83,14 +83,11 @@ impl<'a> Request<'a> {
 
     /// Creates a new `Request` with the given url, desired thumbnail width and
     /// desired thumbnail height.
-    pub fn from_url(url: &'a MxcUri, width: UInt, height: UInt) -> Self {
-        Self {
-            media_id: url.media_id(),
-            server_name: url.server_name(),
-            method: None,
-            width,
-            height,
-            allow_remote: true,
+    pub fn from_url(url: &'a MxcUri, width: UInt, height: UInt) -> Result<Self, Error> {
+        if let (Some(media_id), Some(server_name)) = (url.media_id(), url.server_name()) {
+            Ok(Self { media_id, server_name, method: None, width, height, allow_remote: true })
+        } else {
+            Err(Error::InvalidMxcUri)
         }
     }
 }