+++ ruma-identifiers/src/mxc_uri.rs
@@ -1,75 +1,77 @@
-//! Matrix-spec compliant mxc:// urls.
+//! A URI that should be a Matrix-spec compliant MXC URI.
 
-use std::{convert::TryFrom, fmt, str::FromStr};
+use std::{convert::TryFrom, fmt, num::NonZeroU8};
 
 use ruma_identifiers_validation::mxc_uri::validate;
 
-use crate::{Error, ServerName, ServerNameBox};
+use crate::ServerName;
 
-/// Matrix-spec compliant mxc:// urls.
+/// A URI that should be a Matrix-spec compliant MXC URI.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct MxcUri {
-    server_name: ServerNameBox,
-    media_id: Box<str>,
+    full_uri: Box<str>,
+    slash_idx: Option<NonZeroU8>,
 }
 
 impl MxcUri {
-    /// Returns the media ID of this mxc://.
-    pub fn media_id(&self) -> &str {
-        &self.media_id
+    /// If this is a valid MXC URI, returns the media ID.
+    pub fn media_id(&self) -> Option<&str> {
+        self.slash_idx.map(|idx| &self.full_uri[idx.get() as usize + 1..])
     }
 
-    /// Returns the server name of this mxc://.
-    pub fn server_name(&self) -> &ServerName {
-        &self.server_name
+    /// If this is a valid MXC URI, returns the server name.
+    pub fn server_name(&self) -> Option<&ServerName> {
+        self.slash_idx
+            .map(|idx| <&ServerName>::try_from(&self.full_uri[6..idx.get() as usize]).unwrap())
     }
 
-    fn mxc_uri_fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_fmt(format_args!("mxc://{}/{}", self.server_name, self.media_id))
+    /// Returns if this is a spec-compliant MXC URI.
+    pub fn is_valid(&self) -> bool {
+        self.slash_idx.is_some()
+    }
+
+    /// Create a string slice from this MXC URI.
+    pub fn as_str(&self) -> &str {
+        &self.full_uri
+    }
+
+    /// Create a string slice from this MXC URI.
+    pub fn as_ref(&self) -> &str {
+        self.as_str()
     }
 }
 
 impl fmt::Debug for MxcUri {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.mxc_uri_fmt(f)
+        f.write_str(&self.full_uri)
     }
 }
 
 impl fmt::Display for MxcUri {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.mxc_uri_fmt(f)
+        f.write_str(&self.full_uri)
     }
 }
 
-fn try_from<S>(uri: S) -> Result<MxcUri, Error>
+fn from<S>(uri: S) -> MxcUri
 where
     S: AsRef<str> + Into<Box<str>>,
 {
-    let (media_id, server_name) = validate(uri.as_ref())?;
-    Ok(MxcUri { media_id: media_id.into(), server_name: <ServerNameBox>::try_from(server_name)? })
-}
-
-impl FromStr for MxcUri {
-    type Err = crate::Error;
-
-    fn from_str(uri: &str) -> Result<Self, Self::Err> {
-        try_from(uri)
+    match validate(uri.as_ref()) {
+        Ok(idx) => MxcUri { full_uri: uri.into(), slash_idx: Some(idx) },
+        Err(_) => MxcUri { full_uri: uri.into(), slash_idx: None },
     }
 }
 
-impl TryFrom<&str> for MxcUri {
-    type Error = crate::Error;
-
-    fn try_from(s: &str) -> Result<Self, Self::Error> {
-        try_from(s)
+impl From<&str> for MxcUri {
+    fn from(s: &str) -> Self {
+        from(s)
     }
 }
 
-impl TryFrom<String> for MxcUri {
-    type Error = crate::Error;
-
-    fn try_from(s: String) -> Result<Self, Self::Error> {
-        try_from(s)
+impl From<String> for MxcUri {
+    fn from(s: String) -> Self {
+        from(s)
     }
 }
 
@@ -79,10 +81,7 @@ impl<'de> serde::Deserialize<'de> for MxcUri {
     where
         D: serde::Deserializer<'de>,
     {
-        crate::deserialize_id(
-            deserializer,
-            "Content location represented as a Matrix Content (MXC) URI",
-        )
+        <String>::deserialize(deserializer).map(Into::into)
     }
 }
 
@@ -98,22 +97,52 @@ impl serde::Serialize for MxcUri {
 
 #[cfg(test)]
 mod tests {
-    use std::convert::TryFrom;
+    use matches::assert_matches;
 
     use super::MxcUri;
 
     #[test]
     fn parse_mxc_uri() {
-        assert!(<MxcUri>::try_from("mxc://127.0.0.1/asd32asdfasdsd").is_ok());
+        let mxc = <MxcUri>::from("mxc://127.0.0.1/asd32asdfasdsd");
+
+        assert!(mxc.is_valid());
+        assert_matches!(mxc.media_id(), Some("asd32asdfasdsd"));
+        assert_matches!(mxc.server_name(), Some(name) if name.as_str() == "127.0.0.1")
     }
 
     #[test]
     fn parse_mxc_uri_without_media_id() {
-        assert!(!<MxcUri>::try_from("mxc://127.0.0.1").is_ok());
+        let mxc = <MxcUri>::from("mxc://127.0.0.1");
+
+        assert!(!mxc.is_valid());
+        assert_matches!(mxc.media_id(), None);
+        assert_matches!(mxc.server_name(), None)
     }
 
     #[test]
     fn parse_mxc_uri_without_protocol() {
-        assert!(!<MxcUri>::try_from("127.0.0.1/asd32asdfasdsd").is_ok());
+        assert!(!<MxcUri>::from("127.0.0.1/asd32asdfasdsd").is_valid());
+    }
+
+    #[cfg(feature = "serde")]
+    #[test]
+    fn serialize_mxc_uri() {
+        assert_eq!(
+            serde_json::to_string(&MxcUri::from("mxc://server/1234id"))
+                .expect("Failed to convert MxcUri to JSON."),
+            r#""mxc://server/1234id""#
+        );
+    }
+
+    #[cfg(feature = "serde")]
+    #[test]
+    fn deserialize_mxc_uri() {
+        let mxc = serde_json::from_str::<MxcUri>(r#""mxc://server/1234id""#)
+            .expect("Failed to convert JSON to MxcUri");
+
+        assert_eq!(mxc.as_str(), "mxc://server/1234id");
+        assert!(mxc.is_valid());
+        assert_matches!(mxc.media_id(), Some("1234id"));
+        assert_matches!(mxc.server_name(), Some(name) if name.as_str() == "server")
     }
 }